requirements.txt
# 知识追踪系统依赖包
# 版本锁定，避免兼容性问题

# 核心框架
streamlit>=1.28.0
torch>=1.13.0
numpy>=1.24.0
pandas>=1.3.0
plotly>=5.0.0

# PaddlePaddle 和 PaddleOCR（重要：版本兼容）
paddlepaddle-gpu==2.6.1
paddleocr==3.4.0

# 图像处理
pillow>=9.0.0
opencv-contrib-python==4.10.0.84

# 其他依赖
PyYAML>=6.0
requests>=2.32.0
typing-extensions>=4.12.0
